Reorganize the components and the pages in different package
