+++ frontend/src/App.less
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,300;0,400;0,500;0,700;1,400&display=swap&css');
 
 @import '~antd/dist/antd.less'; // Introduce the official dark less style entry file
-@import 'assets/mainColors';
+@import 'timekeeper-theme.less';
 
 body {
   font-family: 'Montserrat', sans-serif!important;
