--- frontend/src/components/container/Container.less
@@ -1,20 +1,0 @@
-.btn {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    background-color:#e9484e;
-    border-radius:4px;
-    color:#ffffff;
-    span {
-        display: flex;
-        align-items: center;
-    }
-}
-.tk_Menu {
-    margin-top: 70px;
-
-}
-.sider-link {
-    border-color: #FFF;
-    color: #FFF;
-}
