+++ frontend/src/pages/Customer/index.js
@@ -1,11 +1,11 @@
 import React, {useEffect, useState} from 'react'
 import { useLocation } from "react-router-dom";
 import CustomerList from './CustomerList';
-import logo from '../../../img/logo_timekeeper_homepage.png';
+import logo from '../../img/logo_timekeeper_homepage.png';
 import CustomerDetails from "./CustomerDetails";
-import MainPage from "../../MainPage/MainPage";
+import MainPage from "../MainPage/MainPage";
 import NewCustomer from "./NewCustomer";
-import {useAxios} from "../../../utils/hooks";
+import {useAxios} from "../../utils/hooks";
 
 const getCustomerList = (axios, setState) => {
     const fetchData = async () => {
