+++ frontend/src/pages/Home/index.less
@@ -1,4 +1,4 @@
-@import '../../../assets/mainColors.less';
+@import '../../timekeeper-theme.less';
 
 .trigger {
     font-size: 18px;
