+++ frontend/src/pages/Login/index.js
@@ -3,8 +3,8 @@ import { Redirect, withRouter } from 'react-router-dom'
 import { withKeycloak } from '@react-keycloak/web'
 import { Button, Layout, Row, Col, Typography } from "antd";
 
-import Logo from '../../../img/logo.png'
-import LoginBackground from '../../../img/login_Background.png'
+import Logo from '../../img/logo.png'
+import LoginBackground from '../../img/login_Background.png'
 
 import './Login.less'
